savegame: fix overflows in flare count
Resolves #1592. Doesn't get rid of the overflows entirely but makes it
less likely to bump into in a normal playthrough.

+++ src/libtrx/include/libtrx/game/savegame/types.h
@@ -23,7 +23,7 @@ typedef struct {
     uint16_t m16_ammo;
     uint16_t grenade_ammo;
     uint16_t harpoon_ammo;
-    uint8_t flares;
+    uint16_t flares;
 #endif
 
     union {
